requirements.txt
moviepy
imageio[ffmpeg]
Pillow
scenedetect 
streamlit
opencv-python
openai
